Move Graph to new package org.enso.compiler.pass.analyse.alias.graph
