+++ engine/runtime-compiler/src/main/java/org/enso/compiler/pass/analyse/types/LocalBindingsTyping.java
@@ -2,7 +2,7 @@ package org.enso.compiler.pass.analyse.types;
 
 import java.util.HashMap;
 import java.util.Map;
-import org.enso.compiler.pass.analyse.alias.Graph;
+import org.enso.compiler.pass.analyse.alias.graph.Graph;
 
 /**
  * A mapping that maps binding definitions from an alias analysis graph to their inferred types.
